Reworking getting started content
This is the beginning of an edit/rewrite/reorganization of our DevOps Github/Gitbook docs. Bear with me and please help keep me on track. All of my changes will be branched, and none should be merged to master until all are complete or there's an encapsulated set of docs that can be merged without clobbering existing content.
I intend for this Getting Started to be the single entry point for all of the subsequent docs. The primary customer use cases (pathways) will branch off of this. I'm beginning to fill those out.
I'll also be removing nearly all of the content from the main README.md (in https://github.com/pingidentity/pingidentity-devops-getting-started) and reusing that content elsewhere. The main README.me will then essentially contain banner, corporate text, and a link to this Getting Started in the docs directory.